Revert "zigbee: bump to zigbeed-0.5.0"
This reverts commit 5c84acb889a93add600cd86042f9ce6ca9d6a5ce.

Due to hardware problem, the latest firmware (v5.7.4) is required for ARTIK530.
zigbeed-0.5.0 is updated applying v5.7.2 software package for ARTIK530.
ARTIK710 is flashed the old firmware (v5.6.0) now.
However when zigbeed-0.5.0(based on v5.7.4) is applied on ARTIK710,
a critical network connection problem is revealed.
So until fixing this problem, zigbeed-v0.5.0 should not be used due to compatibility issue.
This reverts zigbeed to v0.4.2(based on v5.6.0).

Change-Id: Id3766644ba984487ea8507baa05e88b9f2d6517a
Signed-off-by: Joonghong Park <joonghong.park@samsung.com>